backend/requirements.txt
asgiref==3.8.1
boto3==1.38.19
botocore==1.38.19
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cssselect2==0.8.0
Django==5.2
django-cors-headers==4.7.0
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
django-ratelimit==4.1.0
django-axes==6.1.1
dotenv==0.9.9
et_xmlfile==2.0.0
fonttools==4.58.0
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
openpyxl==3.1.5
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.9
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
s3transfer==0.12.0
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
urllib3==2.4.0
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
